now recognises that + must be a corner

+++ line.js
@@ -22,22 +22,34 @@ const line = (grid) => {
     return false;
 }
 
-const continuePath = (X, Y, grid, turningCoordinates = null) =>{
+const continuePath = (X, Y, grid, disallowedCoordinate = null) =>{
     const surroundingCharsAllowed = getSurroundingCharsAllowed()
     const surroundingSquares = getSurroundingSquares(X,Y,grid)
 
     for (const [relativeDirectionToCurrentSquare, currentSurroundingSymbol] of Object.entries(surroundingSquares)) {
         const symbol = grid[X][Y];
         const allowedSymbols = surroundingCharsAllowed[symbol][relativeDirectionToCurrentSquare];
-        if(allowedSymbols.includes(currentSurroundingSymbol)){
+        const newCoordinates= getNewCoordinates(X,Y,relativeDirectionToCurrentSquare)
+        if(allowedSymbols.includes(currentSurroundingSymbol) && notOnADisAllowedCoordinate(disallowedCoordinate,newCoordinates)){
             if(currentSurroundingSymbol === 'X')
                 return true;
             grid = wipeCurrentSquare(X,Y,grid) //so it doesnt get reprocessed
-            const newCoordinates= getNewCoordinates(X,Y,relativeDirectionToCurrentSquare)
-            return continuePath(newCoordinates.X,newCoordinates.Y,grid);
+            if(currentSurroundingSymbol === "+"){
+                return continuePath(newCoordinates.X,newCoordinates.Y,grid,findDisallowedAxisToFollowPlus({X:X,Y:Y},newCoordinates));
+            }
+            return continuePath(newCoordinates.X,newCoordinates.Y,grid);  //todo save path result so the loop can continue if fails
         }
     }
     return false;
+}
+
+const notOnADisAllowedCoordinate = (disallowedCoordinate,newCoordinates) => {
+    if(!disallowedCoordinate){
+        return true
+    }
+    return !(disallowedCoordinate.X === newCoordinates.X || disallowedCoordinate.Y === newCoordinates.Y);
+
+
 }
 
 const findDisallowedAxisToFollowPlus =(previousCoordinates, plusCoordinates) => {
